+++ src/main/java/com/example/springboot_education/repositories/ActivityLogRepository.java
@@ -1,10 +1,12 @@
 package com.example.springboot_education.repositories;
 
-import com.example.springboot_education.entities.ActivityLog;
-import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import com.example.springboot_education.entities.ActivityLog;
+
 public interface ActivityLogRepository extends JpaRepository<ActivityLog, Integer> {
-    List<ActivityLog> findByUserId(Integer userId);
+    List<ActivityLog> findByUserId(Long userId);
     List<ActivityLog> findByActionType(String actionType);
 }
